Update shell.php
Added an else if statement to system() incase function does not exists and an error gets thrown.
Added passthru() function and another command execution method.
Create a new ruction executeCommand() which accepts the command to run as a parameter. 
Added port and ip parameters for a quick reverse shell. Default port is 443